+++ fabric/src/main/kotlin/com/cobblemonrei/fabric/CobblemonSpawningFabricClient.kt
@@ -17,6 +17,7 @@ class CobblemonSpawningFabricClient : ClientModInitializer {
         ClientTickEvents.END_CLIENT_TICK.register { client ->
             if (client.player != null) {
                 CobblemonSpawningMod.tickReloadCheck()
+                ClientDataReceiver.tick()
             }
         }